+++ modules/Assets/assets/vue-components/field-asset.js
@@ -169,11 +169,11 @@ export default {
     template: /*html*/`
         <div field="asset">
 
-            <div class="kiss-bgcolor-contrast kiss-padding-large kiss-margin-small kiss-align-center kiss-color-muted kiss-position-relative" style="height:150px" v-if="!val">
+            <kiss-card class="kiss-padding-large kiss-margin-small kiss-align-center kiss-color-muted kiss-position-relative" theme="contrast" style="height:150px" v-if="!val">
                 <kiss-svg :src="$base('assets:icon.svg')" width="30" height="30"><canvas width="30" height="30"></canvas></kiss-svg>
                 <div class="kiss-margin-small kiss-size-small">{{ t('No asset selected') }}</div>
                 <a class="kiss-cover" @click="pickAsset()"></a>
-            </div>
+            </kiss-card>
 
             <div class="kiss-margin-small" v-if="val">