+++ modules/Content/views/models/model.php
@@ -102,26 +102,23 @@
                         </tab>
                         <tab :caption="t('Preview')">
 
-                            <kiss-card class="animated fadeIn kiss-padding kiss-align-center kiss-text-caption" theme="bordered contrast" v-if="!model.preview.length">
+                            <kiss-card class="animated fadeIn kiss-padding kiss-align-center kiss-text-caption" theme="contrast" v-if="!model.preview.length">
                                 <div class="kiss-text-bold"><?=t('No content preview urls defined')?></div>
                             </kiss-card>
 
                             <vue-draggable v-model="model.preview" v-if="model.preview.length" handle=".fm-handle">
                                 <template #item="{ element }">
-                                    <kiss-card class="kiss-flex kiss-flex-middle kiss-margin-small">
-                                        <div class="kiss-margin-small-right">
-                                            <icon class="kiss-size-3" :class="{'kiss-color-muted': !(element.name && element.uri)}">visibility</icon>
-                                        </div>
+                                    <kiss-card class="kiss-flex kiss-flex-middle kiss-margin-small kiss-padding-small" gap="small" theme="bordered contrast">
+                                        <a class="fm-handle kiss-color-muted"><icon>drag_handle</icon></a>
                                         <div class="kiss-width-1-4"><input type="text" class="kiss-input kiss-input-small" v-model="element.name" placeholder="<?=t('Name')?>"></div>
-                                        <div class="kiss-margin-small-left kiss-flex-1"><input type="url" class="kiss-input kiss-input-small" v-model="element.uri" placeholder="https://..."></div>
-                                        <a class="kiss-margin-small-left kiss-color-danger" @click="model.preview.splice(model.preview.indexOf(element), 1)"><icon>delete</icon></a>
-                                        <a class="fm-handle kiss-margin-small-left kiss-color-muted"><icon>drag_handle</icon></a>
+                                        <div class="kiss-flex-1"><input type="url" class="kiss-input kiss-input-small" v-model="element.uri" placeholder="https://..."></div>
+                                        <a class="kiss-color-danger" @click="model.preview.splice(model.preview.indexOf(element), 1)"><icon>delete</icon></a>
                                     </kiss-card>
                                 </template>
                             </vue-draggable>
 
-                            <div class="kiss-margin kiss-align-center">
-                                <a class="kiss-size-large" @click="model.preview.push({name:'', uri:''})"><icon>control_point</icon></a>
+                            <div class="kiss-margin-small">
+                                <button type="button" class="kiss-button kiss-button-small" @click="model.preview.push({name:'', uri:''})"><icon class="kiss-margin-small-right">control_point</icon> {{ t('Add url') }}</button>
                             </div>
 
                         </tab>