+++ modules/System/assets/vue-components/fields-manager.js
@@ -160,7 +160,7 @@ let FieldsManager = {
     template: /*html*/`
         <div>
 
-            <kiss-card class="animated fadeIn kiss-padding kiss-align-center kiss-text-caption" theme="bordered contrast" v-if="!fields.length">
+            <kiss-card class="animated fadeIn kiss-padding kiss-align-center kiss-text-caption" theme="contrast" v-if="!fields.length">
                 <div class="kiss-text-bold">{{ t('No fields') }}</div>
             </kiss-card>
 
@@ -192,14 +192,14 @@ let FieldsManager = {
                         </kiss-card>
 
                         <div class="kiss-position-absolute kiss-width-1-3 kiss-align-center kiss-visible-toggle" style="bottom:0;height:20px;left:50%;transform:translateX(-50%) translateY(15%);z-index:5;" v-if="fields.length > 1 && index !== (fields.length - 1)">
-                            <a class="kiss-button kiss-button-small kiss-hidden-hover animated fadeIn faster" :title="t('Add field')" @click="add(element)"><span class="kiss-size-6">+</span></a>
+                            <button type="button" class="kiss-button kiss-button-small kiss-hidden-hover animated fadeIn faster" :title="t('Add field')" @click="add(element)"><span class="kiss-size-6">+</span></button>
                         </div>
                     </div>
                 </template>
             </vue-draggable>
 
-            <div class="kiss-margin kiss-align-center" v-if="fieldTypes">
-                <a class="kiss-size-large" @click="add()"><icon>control_point</icon></a>
+            <div class="kiss-margin-small" v-if="fieldTypes">
+                <button type="button" class="kiss-button kiss-button-small" @click="add()"><icon class="kiss-margin-small-right">control_point</icon> {{ t('Add field') }}</button>
             </div>
 
         </div>